Config options for verbal and visual alerts
Can specify with switches using truthy/falsey values like --verbal-alerts=true
--visual-alerts=false.

The config is persisted to a YAML file, so when running as a daemon, running the
script without any other command will update the config and take effect
immediately.

Changed the interface from #start, to #run, which then runs the given command,
defaulting to start, or just updates the config if no other command is given.

Commands so far are:

* --start
* --stop
* --status
* --report