Fix for appending persistent outages log
Can't actually append YAML, have to load and parse the existing file, and then
truncate and re-write it.

So when starting up, the first thing it now does is attempt to load any existing
outages before it starts writing.

Also updated debug mode in the process so it uses a separate outages log and
leaves the actual outages intact.